// src/types/SearchSummary.ts
export type SearchSummary = {
  ID?: string;
  appName?: string;
  searchName?: string;
  totalResult?: number;
  totalExecResultPercent?: number;
  totalExecResult?: number;
  totalResultPercent?: number;
  totalOutputPercent?: number;
  totalOutput?: number;
  audit?: string;
  vertical?: string;
  jobName?: string;
};